+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @atomicsmash/cli
 
+## 8.1.0-beta.0
+
+### Minor Changes
+
+- [#382](https://github.com/AtomicSmash/packages/pull/382) [`33a7304`](https://github.com/AtomicSmash/packages/commit/33a73044bf3dcc9f3d745bef4c0c1f6387f01432) Thanks [@mikeybinns](https://github.com/mikeybinns)! - Allow customising of scss aliases with scssAliases.config.ts file
+
+- [#382](https://github.com/AtomicSmash/packages/pull/382) [`33a7304`](https://github.com/AtomicSmash/packages/commit/33a73044bf3dcc9f3d745bef4c0c1f6387f01432) Thanks [@mikeybinns](https://github.com/mikeybinns)! - Add Launchpad alias to default scss aliases
+
 ## 8.0.0
 
 ### Major Changes