+++ packages/compiler/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @atomicsmash/compiler
 
+## 1.1.0-beta.2
+
+### Minor Changes
+
+- [#382](https://github.com/AtomicSmash/packages/pull/382) [`33a7304`](https://github.com/AtomicSmash/packages/commit/33a73044bf3dcc9f3d745bef4c0c1f6387f01432) Thanks [@mikeybinns](https://github.com/mikeybinns)! - Allow customising of scss aliases with scssAliases.config.ts file
+
+- [#382](https://github.com/AtomicSmash/packages/pull/382) [`33a7304`](https://github.com/AtomicSmash/packages/commit/33a73044bf3dcc9f3d745bef4c0c1f6387f01432) Thanks [@mikeybinns](https://github.com/mikeybinns)! - Add Launchpad alias to default scss aliases
+
 ## 1.0.1-beta.1
 
 ### Patch Changes